+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]## [0.1.2.4b4]
 ### Changed
 ### Added
+* Added uv support
 ### Fixed
 - KpiPageTemplate.get_kpi_grid fixed bug where column show NaN if scenario name contains a dot.