+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]## [0.1.2.4b4]
 ### Changed
+* DashApp. Removed diskcache import
 ### Added
 * Added uv support
 ### Fixed