+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]## [0.1.2.4b4]
+
+## [0.1.2.4b4] - 2026-09-14
 ### Changed
 * DashApp. Removed diskcache import
 ### Added